+++ internal/legacy_files/doc.go
@@ -1,5 +1,0 @@
-/*
-Package swag converts Go annotations to Swagger Documentation 2.0.
-See https://github.com/griffnb/core-swag for more information about swag.
-*/
-package swag // import "github.com/griffnb/core-swag"